requirements.txt
Django==5.0.3
djangorestframework==3.15.0
psycopg2-binary==2.9.10
pandas==2.2.0
openpyxl==3.1.5
celery==5.4.0
redis==5.0.1
python-dotenv==1.0.1
gunicorn==22.0.0
